moving the test folder outside from src folder
