+++ client-side/package.json
@@ -10,13 +10,6 @@
     "@mui/material": "^5.16.0",
     "@mui/styled-engine-sc": "^6.0.0-alpha.18",
     "@mui/x-charts": "^7.10.0",
-    "@mui/x-data-grid": "^7.10.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "html2canvas": "^1.4.1",
-    "html2pdf.js": "^0.10.2",
-    "jspdf": "^2.5.1",
-    "jspdf-autotable": "^3.8.2",
     "@mui/x-data-grid": "^7.9.0",
     "@reduxjs/toolkit": "^2.2.6",
     "@testing-library/jest-dom": "^5.17.0",
@@ -24,6 +17,11 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/redux": "^3.6.31",
     "axios": "^1.7.2",
+    "html2canvas": "^1.4.1",
+    "html2pdf.js": "^0.10.2",
+    "jest": "^29.7.0",
+    "jspdf": "^2.5.1",
+    "jspdf-autotable": "^3.8.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
