+++ client-side/test/loader.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Loader from '../stories/loader/loader.jsx'; 
+import Loader from '../src/stories/loader/loader.jsx'; 
 
 test('renders Loader component with default class', () => {
   render(<Loader />);
