+++ client-side/test/popup.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import SimplePopup from '../stories/popup/popup';
+import SimplePopup from '../src/stories/popup/popup';
 
 test('renders the SimplePopup component and toggles the popup', () => {
   const customMessage = 'Hello, this is a custom message!';
